Added password management
I added password function to check user's password.

+++ webapp/audio_file_remove.php
@@ -14,6 +14,7 @@
 $file_id   = $_REQUEST['file_id'];
 $name_orig = $_REQUEST['name_orig'];
 $name_save = $_REQUEST['name_save'];
+$password  = $_POST['pass'];
 
 echo ("<br>");
 echo ("<br>");
@@ -31,27 +32,44 @@ if (!$db_conn) {
     die("Connection failed: " . mysqli_db_connect_error());
 }
 
-// sql to delete a record
-$sql = "DELETE FROM upload_file WHERE file_id='$file_id'";
+// compare your password and database password.
+// echo ("[DEBUG] select count(*) FROM upload_file WHERE file_id = '$file_id' and password = '$password' <br>");
+$sql = "select count(*) FROM upload_file WHERE file_id = '$file_id' and password = '$password'";
+$result_count=mysqli_query($db_conn, $sql);
+$result_row=mysqli_fetch_row($result_count);
+$total_row = $result_row[0];
 
-if (mysqli_query($db_conn, $sql)) {
-    echo "<font color=red><b>축하합니다.</font><b> 선택하신 음성파일을 성공적으로 삭제되었습니다.";
-} else {
-    echo "<font color=red><b>죄송합니다.</font><b> 선택하신 음성파일을 삭제하지 못하였습니다." . mysqli_error($db_conn);
+//echo ("[DEBUG] matched count is $total_row .<br>");
+//die ("[DEBUG] Let's do debugging");
+
+// if $total_row is 0, do not remove audio file and db data.
+if ($total_row <=0)
+    echo ("<font color=red>죄송합니다.</font> 비밀번호가 틀립니다. 올바르게 비밀번호를 입력하여 주세요.");
+else {
+    // sql to delete a record
+    $sql = "DELETE FROM upload_file WHERE file_id='$file_id'";
+    
+    if (mysqli_query($db_conn, $sql)) {
+        echo "<font color=red><b>축하합니다.</font><b> 선택하신 음성파일을 성공적으로 삭제되었습니다.";
+    } else {
+        echo "<font color=red><b>죄송합니다.</font><b> 선택하신 음성파일을 삭제하지 못하였습니다." . mysqli_error($db_conn);
+    }
+    
+    // get folder name from $name_save (e.g., 201807231430_a872sadj29w891092d.m4a)
+    $audio_folder = substr($name_save,0,8);
+    echo ("<br>");
+    echo ("<br>");
+    echo ("오디오 파일의 날짜 폴더명은 ".$audio_folder."입니다.");
+    
+    // Remove audio files in the specified directory
+    unlink("audio/".$audio_folder."/".$name_save);
+    if (file_exists("audio/".$audio_folder."/".$name_save.".txt")){
+        unlink("audio/".$audio_folder."/".$name_save.".txt");
+    }
+    mysqli_close($db_conn);
 }
-
-// get folder name from $name_save (e.g., 201807231430_a872sadj29w891092d.m4a)
-$audio_folder = substr($name_save,0,8);
-echo ("<br>");
-echo ("<br>");
-echo ("오디오 파일의 폴더명은 ".$audio_folder."입니다.");
-
-// Remove audito file in the specified directory
-unlink("audio/".$audio_folder."/".$name_save);
-unlink("audio/".$audio_folder."/".$name_save.".txt");
-
-mysqli_close($db_conn);
 ?>
+
 <br><br>
 <a href="./audio_file_list.php"><img src=./images/file-list.png alt="오디오 파일리스트로 이동하기" title="오디오 파일리스트로 이동하기" border=0 width=50 height=50></img></a>
 </body>
